+++ common/src/main/kotlin/com/pabi/common/exception/ServerException.kt
@@ -7,7 +7,7 @@ sealed class ServerException(
 
 data class InvalidTokenException(
     override val message: String = "유효하지 않은 토큰 입니다."
-) : ServerException(200, message)
+) : ServerException(401, message)
 
 data class DuplicateUserNickNameException(
     override val message: String = "이미 존재 하는 닉네임 입니다.",
