+++ user/domain/src/main/kotlin/com/pabi/user/domain/service/TokenService.kt
@@ -1,31 +1,17 @@
 package com.pabi.user.domain.service
 
-import com.pabi.common.exception.InvalidTokenException
 import com.pabi.common.jwt.TokenProvider
+import com.pabi.common.response.Token
 import org.springframework.stereotype.Service
-import org.springframework.util.StringUtils
 
 @Service
 class TokenService(
     private val tokenProvider: TokenProvider,
 ) {
 
-    fun tokenReissue(accessToken: String, refreshToken: String): String {
-        val access = resolveToken(accessToken)
-        val refresh = resolveToken(refreshToken)
+    fun tokenReissue(accessToken: String, refreshToken: String): Token {
+        val access = tokenProvider.resolveToken(accessToken)
+        val refresh = tokenProvider.resolveToken(refreshToken)
         return tokenProvider.tokenReissue(access, refresh)
-    }
-
-    fun resolveToken(token: String): String {
-
-        if (!StringUtils.hasText(token)) {
-            throw InvalidTokenException()
-        }
-
-        if (!token.startsWith("Bearer ")) {
-            throw InvalidTokenException()
-        }
-
-        return token.substring(7)
     }
 }
