+++ user/presentation/src/main/kotlin/com/pabi/user/presentation/TokenReissueController.kt
@@ -3,6 +3,7 @@ package com.pabi.user.presentation
 import com.pabi.common.jwt.JwtFilter.Companion.AUTHORIZATION_HEADER
 import com.pabi.common.jwt.JwtFilter.Companion.REFRESH_TOKEN_HEADER
 import com.pabi.common.response.CommonResponse
+import com.pabi.common.response.Token
 import com.pabi.user.application.TokenFacade
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestHeader
@@ -19,7 +20,7 @@ class TokenReissueController(
     fun tokenReissue(
         @RequestHeader(AUTHORIZATION_HEADER) accessToken: String,
         @RequestHeader(REFRESH_TOKEN_HEADER) refreshToken: String,
-    ): CommonResponse<String> {
+    ): CommonResponse<Token> {
         val token = tokenFacade.tokenReissue(accessToken, refreshToken)
         return CommonResponse.success(token, "토큰이 재발급 되었습니다.");
     }
